Final/requirements.txt
openai
python-dotenv
numpy==2.2.3
opencv-python==4.11.0.86
SpeechRecognition==3.14.1
typing_extensions==4.12.2
requests